moved helper functions to own file, added more documentation to transform
